server/requirements.txt
# FastAPI server dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# TripoSR dependencies
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.1.0
omegaconf==2.3.0
einops==0.7.0
transformers==4.35.0
trimesh>=4.0.5
rembg
huggingface-hub
imageio[ffmpeg]
